requirements.txt
aiosqlite==0.21.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
black==25.11.0
certifi==2025.11.12
click==8.3.1
colorama==0.4.6
fastapi==0.121.3
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
mypy_extensions==1.1.0
nats-py==2.12.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.0
playwright==1.56.0
pydantic==2.12.4
pydantic_core==2.41.5
pyee==13.0.0
pytokens==0.3.0
ruff==0.14.5
sniffio==1.3.1
SQLAlchemy==2.0.44
sqlmodel==0.0.27
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
websockets==15.0.1